backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
alembic
